[#11024] DocDB: Enable -Wshorten-64-to-32 flag
Summary:
This diff enables -Wshorten-64-to-32 flag and fixes all places where this warning appear in our code.
The following strategy was used:
1) Minimize amount of casts, i.e. changing used type is more preferable.
2) For places where such cast is necessary, for instance because of 3rd party library API, use newly introduced `narrow_cast`.
`narrow_cast` expects that casted values left unchanged. I.e. if we cast 64 bit number to 32 bit number it expects that actual value fits into 32 bits.

Test Plan: Jenkins

Reviewers: mbautin, mpolitov

Reviewed By: mpolitov

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D14653

+++ ent/src/yb/integration-tests/encryption-test.cc
@@ -60,11 +60,11 @@ class EncryptionTest : public YBTableTestBase, public testing::WithParamInterfac
 
   bool enable_ysql() override { return false; }
 
-  int num_tablet_servers() override {
+  size_t num_tablet_servers() override {
     return 3;
   }
 
-  int num_masters() override {
+  size_t num_masters() override {
     return 3;
   }
