+++ src/yb/consensus/quorum_util.cc
@@ -175,7 +175,7 @@ int CountMemberType(const RaftConfigPB& config, const PeerMemberType member_type
   return count;
 }
 
-int MajoritySize(int num_voters) {
+size_t MajoritySize(size_t num_voters) {
   DCHECK_GE(num_voters, 1);
   return (num_voters / 2) + 1;
 }
