+++ src/yb/gutil/spinlock_internal.cc
@@ -112,7 +112,7 @@ static int SuggestedDelayNS(int loop) {
   // Mean is exponential in loop for first 32 iterations, then 8ms.
   // The futex path multiplies this by 16, since we expect explicit wakeups
   // almost always on that path.
-  return r >> (44 - (loop >> 3));
+  return static_cast<int>(r >> (44 - (loop >> 3)));
 #else
   static Atomic32 rand;
   uint32 r = base::subtle::NoBarrier_Load(&rand);
