+++ src/yb/gutil/valgrind.h
@@ -3652,7 +3652,7 @@ typedef
    running under Valgrind which is running under another Valgrind,
    etc. */
 #define RUNNING_ON_VALGRIND  __extension__                        \
-   ({unsigned int _qzz_res;                                       \
+   ({unsigned long long _qzz_res;                                 \
     VALGRIND_DO_CLIENT_REQUEST(_qzz_res, 0 /* if not */,          \
                                VG_USERREQ__RUNNING_ON_VALGRIND,   \
                                0, 0, 0, 0, 0);                    \
